+++ src/main/java/com/pixelandtag/cmp/ejb/CMPResourceBeanRemote.java
@@ -26,7 +26,7 @@ import com.pixelandtag.web.beans.RequestObject;
 
 public interface CMPResourceBeanRemote extends BaseEntityI {
 	
-	public boolean updateMessageInQueue(long cp_tx_id, BillingStatus billstatus) throws Exception;
+	public boolean updateMessageInQueue(BigInteger cp_tx_id, BillingStatus billstatus) throws Exception;
 	
 	public <T> Collection<T> find(Class<T> entityClass,Map<String, Object> criteria, int start, int end)  throws Exception ;
